fix: empty iss return
